pyproject.toml
[project]
name = "kisan-voice-bot"
version = "1.0.0"
description = "Multilingual agricultural assistance bot for Indian farmers"
requires-python = ">=3.11, <3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "openai-whisper>=20231117",
    "torch>=2.1.2",
    "torchaudio>=2.1.2",
    "ollama>=0.1.6",
    "coqui-tts>=0.24.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.3.1",
    "google-genai>=1.0.0",
    "elevenlabs>=2.29.0",
    "edge-tts>=6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = []
